refactor(telemetry): merge OTLP handlers from otlp.py, delete otlp.py
