+++ observal-server/main.py
@@ -36,12 +36,12 @@ from api.routes.feedback import router as feedback_router
 from api.routes.hook import router as hook_router
 from api.routes.jwks import router as jwks_router
 from api.routes.mcp import router as mcp_router
-from api.routes.otlp import router as otlp_router
 from api.routes.prompt import router as prompt_router
 from api.routes.review import router as review_router
 from api.routes.sandbox import router as sandbox_router
 from api.routes.sessions import router as sessions_router
 from api.routes.skill import router as skill_router
+from api.routes.telemetry import otlp_router
 from api.routes.telemetry import router as telemetry_router
 from config import settings
 from database import engine
